Iteration #4 Lots initial flow - Dev + Bid start

+++ db/migrate/20180806095133_create_bids.rb
@@ -2,7 +2,7 @@ class CreateBids < ActiveRecord::Migration[5.2]
   def change
     create_table :bids do |t|
 
-      t.decimal :proposed_price
+      t.float :proposed_price
       t.datetime :created_at
 
       t.belongs_to :lot, foreign_key: true, index: true
